Move if-down script out of netplan/tools down/off path.
This changed is intended to rework how the if-down script is handled
for removal of the routes/tables/rules created by the charm.
Currently, this cleanup script is placed in either net-tools or
networkd-dispatch etc directories if-down.d/off.d respectively. This placement
causes issues as any interface powering down invokes these scripts, removing
the configured routes as a result. Here, we keep the cleanup script but
avoid creating that symlink in those if-down.d/off.d directories.